requirements.txt
langchain==0.3.15
langchain-core==0.3.31
langgraph==0.2.67
langchain-community==0.3.15
langchain-openai==0.3.1
nicegui==2.10.1
python-dotenv==1.0.1
pydantic==2.10.5
upstash-vector==0.7.0
google-cloud-speech==2.30.0
pydub==0.25.1
supabase==2.12.0
pandas==2.2.3
Authlib